refactor(admin): extract Variables section from AppFormEditor (#1781)
Continues the incremental split proposed in #1781 (and started in #1942
for Upload Configuration). Moves the Variables Configuration card and
its six immutable-update handlers (handleVariableChange,
handleVariablePredefinedValueChange, addPredefinedValue,
removePredefinedValue, addVariable, removeVariable) into a
self-contained VariablesSection component, since none of them touch
anything outside app.variables.

AppFormEditor.jsx shrinks from 2902 to 2618 lines. No behavior change
intended - JSX and handler logic ported verbatim.
